% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ics_tests.R
\name{calculate_test_statistic}
\alias{calculate_test_statistic}
\title{Helper function to calculate test statistic}
\usage{
calculate_test_statistic(wYvec, Zvec, cluster_sizes, covariates, adjust_size)
}
\description{
Helper function to calculate test statistic
}
\keyword{internal}
